+++ pygeos/creation.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from . import Geometry, GeometryType, lib
-from ._geometry import collections_1d, polygons_1d, simple_geometries_1d
+from ._geometry import collections_1d, simple_geometries_1d
 from .decorators import multithreading_enabled
 
 __all__ = [
@@ -46,12 +46,19 @@ def points(coords, y=None, z=None, indices=None, **kwargs):
     indices : array_like or None
         Indices into the target array where input coordinates belong. If
         provided, the coords should be 2D with shape (N, 2) or (N, 3) and
-        indices should be 1D with shape (N,). Missing indices will give None
-        values in the output array.
+        indices should be an array of shape (N,) with integers in increasing
+        order. Missing indices will result in a ValueError.
     **kwargs
         For other keyword-only arguments, see the
         `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
         Ignored if ``indices`` is provided.
+
+    Examples
+    --------
+    >>> points([[0, 1], [4, 5]]).tolist()
+    [<pygeos.Geometry POINT (0 1)>, <pygeos.Geometry POINT (4 5)>]
+    >>> points([0, 1, 2])
+    <pygeos.Geometry POINT Z (0 1 2)>
     """
     coords = _xyz_to_coords(coords, y, z)
     if indices is None:
@@ -77,12 +84,19 @@ def linestrings(coords, y=None, z=None, indices=None, **kwargs):
     indices : array_like or None
         Indices into the target array where input coordinates belong. If
         provided, the coords should be 2D with shape (N, 2) or (N, 3) and
-        indices should be 1D with shape (N,). Missing indices will give None
-        values in the output array.
+        indices should be an array of shape (N,) with integers in increasing
+        order. Missing indices will result in a ValueError.
     **kwargs
         For other keyword-only arguments, see the
         `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
         Ignored if ``indices`` is provided.
+
+    Examples
+    --------
+    >>> linestrings([[[0, 1], [4, 5]], [[2, 3], [5, 6]]]).tolist()
+    [<pygeos.Geometry LINESTRING (0 1, 4 5)>, <pygeos.Geometry LINESTRING (2 3, 5 6)>]
+    >>> linestrings([[0, 1], [4, 5], [2, 3], [5, 6], [7, 8]], indices=[0, 0, 1, 1, 1]).tolist()
+    [<pygeos.Geometry LINESTRING (0 1, 4 5)>, <pygeos.Geometry LINESTRING (2 3, 5 6, 7 8)>]
     """
     coords = _xyz_to_coords(coords, y, z)
     if indices is None:
@@ -110,12 +124,23 @@ def linearrings(coords, y=None, z=None, indices=None, **kwargs):
     indices : array_like or None
         Indices into the target array where input coordinates belong. If
         provided, the coords should be 2D with shape (N, 2) or (N, 3) and
-        indices should be 1D with shape (N,). Missing indices will give None
-        values in the output array.
+        indices should be an array of shape (N,) with integers in increasing
+        order. Missing indices will result in a ValueError.
     **kwargs
         For other keyword-only arguments, see the
         `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
         Ignored if ``indices`` is provided.
+
+    See also
+    --------
+    linestrings
+
+    Examples
+    --------
+    >>> linearrings([[0, 0], [0, 1], [1, 1], [0, 0]])
+    <pygeos.Geometry LINEARRING (0 0, 0 1, 1 1, 0 0)>
+    >>> linearrings([[0, 0], [0, 1], [1, 1]])
+    <pygeos.Geometry LINEARRING (0 0, 0 1, 1 1, 0 0)>
     """
     coords = _xyz_to_coords(coords, y, z)
     if indices is None:
@@ -125,34 +150,86 @@ def linearrings(coords, y=None, z=None, indices=None, **kwargs):
 
 
 @multithreading_enabled
-def polygons(shells, holes=None, indices=None, **kwargs):
+def polygons(geometries, holes=None, indices=None, **kwargs):
     """Create an array of polygons.
+    and the holes; the first geometry for each index is the outer shell and all subsequent geometries in that index are the holes...
+        Parameters
+        ----------
+        geometries : array_like
+            An array of linearrings or coordinates (see linearrings).
+            Unless ``indices`` are given (see description below), this
+            include the outer shells only. The ``holes`` argument should be used
+            to create polygons with holes.
+        holes : array_like or None
+            An array of lists of linearrings that constitute holes for each shell.
+            Not to be used in combination with ``indices``.
+        indices : array_like or None
+            Indices into the target array where input geometries belong. If
+            provided, the holes are expected to be present inside ``geometries``;
+            the first geometry for each index is the outer shell
+            and all subsequent geometries in that index are the holes.
+            Both geometries and indices should be 1D and have matching sizes.
+            Indices should be in increasing order. Missing indices will result
+            in a ValueError.
+        **kwargs
+            For other keyword-only arguments, see the
+            `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
+            Ignored if ``indices`` is provided.
 
-    Parameters
-    ----------
-    shell : array_like
-        An array of linearrings that constitute the out shell of the polygons.
-        Coordinates can also be passed, see linearrings.
-    holes : array_like or None
-        An array of lists of linearrings that constitute holes for each shell.
-    indices : array_like or None
-        Indices into the shells array where input holes belong. If
-        provided, shells, holes, and indices should all be 1D and the size
-        of holes must equal the size of indices.
-    **kwargs
-        For other keyword-only arguments, see the
-        `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
-        Ignored if ``indices`` is provided.
+        Examples
+        --------
+        Polygons are constructed from rings:
+
+        >>> ring_1 = linearrings([[0, 0], [0, 10], [10, 10], [10, 0]])
+        >>> ring_2 = linearrings([[2, 6], [2, 7], [3, 7], [3, 6]])
+        >>> polygons([ring_1, ring_2])[0]
+        <pygeos.Geometry POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))>
+        >>> polygons([ring_1, ring_2])[1]
+        <pygeos.Geometry POLYGON ((2 6, 2 7, 3 7, 3 6, 2 6))>
+
+        Or from coordinates directly:
+
+        >>> polygons([[0, 0], [0, 10], [10, 10], [10, 0]])
+        <pygeos.Geometry POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))>
+
+        Adding holes can be done using the `holes` keyword argument:
+
+        >>> polygons(ring_1, holes=[ring_2])
+        <pygeos.Geometry POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0), (2 6, 2 7, 3 7, 3 6...>
+
+        Or using the `indices` argument:
+
+        >>> polygons([ring_1, ring_2], indices=[0, 1])[0]
+        <pygeos.Geometry POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))>
+        >>> polygons([ring_1, ring_2], indices=[0, 1])[1]
+        <pygeos.Geometry POLYGON ((2 6, 2 7, 3 7, 3 6, 2 6))>
+        >>> polygons([ring_1, ring_2], indices=[0, 0])[0]
+        <pygeos.Geometry POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0), (2 6, 2 7, 3 7, 3 6...>
+
+        Missing input values (`None`) are ignored and may result in an
+        empty polygon:
+
+        >>> polygons(None)
+        <pygeos.Geometry POLYGON EMPTY>
+        >>> polygons(ring_1, holes=[None])
+        <pygeos.Geometry POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))>
+        >>> polygons([ring_1, None], indices=[0, 0])[0]
+        <pygeos.Geometry POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))>
     """
-    shells = np.asarray(shells)
-    if not isinstance(shells, Geometry) and np.issubdtype(shells.dtype, np.number):
-        shells = linearrings(shells)
+    geometries = np.asarray(geometries)
+    if not isinstance(geometries, Geometry) and np.issubdtype(
+        geometries.dtype, np.number
+    ):
+        geometries = linearrings(geometries)
 
-    if holes is None and indices is not None:
-        raise TypeError("Indices provided without a holes array.")
-    elif holes is None:
+    if indices is not None:
+        if holes is not None:
+            raise TypeError("Cannot specify separate holes array when using indices.")
+        return collections_1d(geometries, indices, GeometryType.POLYGON)
+
+    if holes is None:
         # no holes provided: initialize an empty holes array matching shells
-        shape = shells.shape + (0,) if isinstance(shells, np.ndarray) else (0,)
+        shape = geometries.shape + (0,) if isinstance(geometries, np.ndarray) else (0,)
         holes = np.empty(shape, dtype=object)
     else:
         holes = np.asarray(holes)
@@ -160,10 +237,7 @@ def polygons(shells, holes=None, indices=None, **kwargs):
         if np.issubdtype(holes.dtype, np.number):
             holes = linearrings(holes)
 
-    if indices is None:
-        return lib.polygons(shells, holes, **kwargs)
-    else:
-        return polygons_1d(shells, holes, indices)
+    return lib.polygons(geometries, holes, **kwargs)
 
 
 @multithreading_enabled
@@ -207,11 +281,44 @@ def multipoints(geometries, indices=None, **kwargs):
     indices : array_like or None
         Indices into the target array where input geometries belong. If
         provided, both geometries and indices should be 1D and have matching
-        sizes.
+        sizes. Indices should be in increasing order. Missing indices will result
+        in a ValueError.
     **kwargs
         For other keyword-only arguments, see the
         `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
         Ignored if ``indices`` is provided.
+
+    Examples
+    --------
+    Multipoints are constructed from points:
+
+    >>> point_1 = points([1, 1])
+    >>> point_2 = points([2, 2])
+    >>> multipoints([point_1, point_2])
+    <pygeos.Geometry MULTIPOINT (1 1, 2 2)>
+    >>> multipoints([[point_1, point_2], [point_2, None]]).tolist()
+    [<pygeos.Geometry MULTIPOINT (1 1, 2 2)>, <pygeos.Geometry MULTIPOINT (2 2)>]
+
+    Or from coordinates directly:
+
+    >>> multipoints([[0, 0], [2, 2], [3, 3]])
+    <pygeos.Geometry MULTIPOINT (0 0, 2 2, 3 3)>
+
+    Multiple multipoints of different sizes can be constructed efficiently using the
+    `indices` keyword argument:
+
+    >>> multipoints([point_1, point_2, point_2], indices=[0, 0, 1]).tolist()
+    [<pygeos.Geometry MULTIPOINT (1 1, 2 2)>, <pygeos.Geometry MULTIPOINT (2 2)>]
+
+    Missing input values (`None`) are ignored and may result in an
+    empty multipoint:
+
+    >>> multipoints([None])
+    <pygeos.Geometry MULTIPOINT EMPTY>
+    >>> multipoints([point_1, None], indices=[0, 0]).tolist()
+    [<pygeos.Geometry MULTIPOINT (1 1)>]
+    >>> multipoints([point_1, None], indices=[0, 1]).tolist()
+    [<pygeos.Geometry MULTIPOINT (1 1)>, <pygeos.Geometry MULTIPOINT EMPTY>]
     """
     typ = GeometryType.MULTIPOINT
     geometries = np.asarray(geometries)
@@ -236,11 +343,16 @@ def multilinestrings(geometries, indices=None, **kwargs):
     indices : array_like or None
         Indices into the target array where input geometries belong. If
         provided, both geometries and indices should be 1D and have matching
-        sizes.
+        sizes. Indices should be in increasing order. Missing indices will result
+        in a ValueError.
     **kwargs
         For other keyword-only arguments, see the
         `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
         Ignored if ``indices`` is provided.
+
+    See also
+    --------
+    multipoints
     """
     typ = GeometryType.MULTILINESTRING
     geometries = np.asarray(geometries)
@@ -266,11 +378,16 @@ def multipolygons(geometries, indices=None, **kwargs):
     indices : array_like or None
         Indices into the target array where input geometries belong. If
         provided, both geometries and indices should be 1D and have matching
-        sizes.
+        sizes. Indices should be in increasing order. Missing indices will result
+        in a ValueError.
     **kwargs
         For other keyword-only arguments, see the
         `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
         Ignored if ``indices`` is provided.
+
+    See also
+    --------
+    multipoints
     """
     typ = GeometryType.MULTIPOLYGON
     geometries = np.asarray(geometries)
@@ -295,11 +412,16 @@ def geometrycollections(geometries, indices=None, **kwargs):
     indices : array_like or None
         Indices into the target array where input geometries belong. If
         provided, both geometries and indices should be 1D and have matching
-        sizes.
+        sizes. Indices should be in increasing order. Missing indices will result
+        in a ValueError.
     **kwargs
         For other keyword-only arguments, see the
         `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
         Ignored if ``indices`` is provided.
+
+    See also
+    --------
+    multipoints
     """
     typ = GeometryType.GEOMETRYCOLLECTION
     if indices is None: