+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ Version 0.10 (unreleased)
 * Addition of ``nearest`` and ``nearest_all`` functions to ``STRtree`` for
   GEOS >= 3.6 to find the nearest neighbors (#272).
 * Enable bulk construction of geometries with different number of coordinates
-  by optionally taking index arrays in all creation functions (#230, #322, #326).
+  by optionally taking index arrays in all creation functions (#230, #322, #326, #346).
 * Released the GIL in all geometry creation functions (#310, #326).
 * Added the option to return the geometry index in ``get_coordinates`` (#318).
 * Updated ``box`` ufunc to use internal C function for creating polygon
@@ -31,7 +31,9 @@ Version 0.10 (unreleased)
   optionally return invalid WKB geometries as ``None``.
 * Removed the (internal) function ``lib.polygons_without_holes`` and renamed
   ``lib.polygons_with_holes`` to ``lib.polygons`` (#326).
+* ``polygons`` will now return an empty polygon for `None` inputs (#346).
 * Removed compatibility with Python 3.5 (#341).
+
 
 **Added GEOS functions**